Refactored feedback submission to use documentId as a long type and update related components for consistency

+++ search_engine_api/src/main/java/com/suyash/search_engine_api/feedback/Feedback.java
@@ -31,7 +31,7 @@ public class Feedback {
 
     private String query;
 
-    private int documentId;
+    private long documentId;
 
     private boolean isRelevant;
